refactor(browser): split main.rs into focused modules
Move History, Tab, DocumentLoader, TabStrip, Toolbar, and StatusBar
into their own files. Expose a public HistoryNav delegation trait to
bridge the private trait generated by #[store] across module boundaries.
