+++ apps/browser/src/history.rs
@@ -1,0 +1,109 @@
+use blitz_traits::navigation::{NavigationOptions, NavigationProvider};
+use blitz_traits::net::{Request, Url};
+use dioxus_native::prelude::*;
+
+pub type SyncStore<T> = Store<T, CopyValue<T, SyncStorage>>;
+
+#[derive(Store)]
+pub struct History {
+    urls: Vec<Request>,
+    current: usize,
+}
+
+impl History {
+    pub fn new(initial_url: Url) -> Self {
+        Self {
+            urls: vec![Request::get(initial_url)],
+            current: 0,
+        }
+    }
+}
+
+#[store]
+impl<Lens> Store<History, Lens> {
+    fn current_idx(&self) -> usize {
+        *self.current().read()
+    }
+
+    fn current_url(&self) -> impl Readable<Target = Request> {
+        self.urls().get(self.current_idx()).unwrap()
+    }
+
+    fn has_back(&self) -> bool {
+        self.current_idx() > 0
+    }
+
+    fn has_forward(&self) -> bool {
+        self.current_idx() < self.urls().len() - 1
+    }
+
+    fn go_back(&mut self) {
+        if self.has_back() {
+            *self.current().write() -= 1;
+        }
+    }
+
+    fn go_forward(&mut self) {
+        if self.has_forward() {
+            *self.current().write() += 1;
+        }
+    }
+
+    fn navigate(&self, req: Request)
+    where
+        Lens: Writable,
+    {
+        let idx = self.current_idx();
+        self.urls().write().truncate(idx + 1);
+        self.urls().push(req);
+        *self.current().write() += 1;
+    }
+}
+
+/// Public extension trait exposing history navigation for use across modules.
+/// Delegates to the private `HistoryStoreImplExt` generated by `#[store]`,
+/// which is only accessible within this module.
+pub trait HistoryNav {
+    fn current_url(&self) -> impl Readable<Target = Request>;
+    fn has_back(&self) -> bool;
+    fn has_forward(&self) -> bool;
+    fn go_back(&mut self);
+    fn go_forward(&mut self);
+    fn navigate(&self, req: Request);
+}
+
+impl HistoryNav for SyncStore<History> {
+    fn current_url(&self) -> impl Readable<Target = Request> {
+        HistoryStoreImplExt::current_url(self)
+    }
+
+    fn has_back(&self) -> bool {
+        HistoryStoreImplExt::has_back(self)
+    }
+
+    fn has_forward(&self) -> bool {
+        HistoryStoreImplExt::has_forward(self)
+    }
+
+    fn go_back(&mut self) {
+        HistoryStoreImplExt::go_back(self)
+    }
+
+    fn go_forward(&mut self) {
+        HistoryStoreImplExt::go_forward(self)
+    }
+
+    fn navigate(&self, req: Request) {
+        HistoryStoreImplExt::navigate(self, req)
+    }
+}
+
+pub struct BrowserNavProvider {
+    pub history: SyncStore<History>,
+}
+
+impl NavigationProvider for BrowserNavProvider {
+    fn navigate_to(&self, options: NavigationOptions) {
+        HistoryNav::navigate(&self.history, options.into_request());
+    }
+}
